[not verified] add visibility settings to block inspector
